+++ app/layout/scenarios/edit/features/set-up/add/intersect/toolbar/index.ts
@@ -1,0 +1,1 @@
+export { default } from './component';
