+++ app/layout/scenarios/edit/features/set-up/targets/component.tsx
@@ -1,313 +1,106 @@
-import React, { useCallback, useMemo, useState } from 'react';
+import React, { useEffect } from 'react';
 
-import { Form as FormRFF, Field as FieldRFF } from 'react-final-form';
+import { useSelector, useDispatch } from 'react-redux';
 
 import { useRouter } from 'next/router';
 
-import cx from 'classnames';
-import { ScenarioSidebarTabs } from 'utils/tabs';
-import { mergeScenarioStatusMetaData } from 'utils/utils-scenarios';
+import { getScenarioEditSlice } from 'store/slices/scenarios/edit';
 
-import {
-  useSaveSelectedFeatures, useSelectedFeatures, useTargetedFeatures,
-} from 'hooks/features';
-import { useSaveScenario, useScenario } from 'hooks/scenarios';
+import { motion } from 'framer-motion';
+import { ScenarioSidebarSubTabs, ScenarioSidebarTabs } from 'utils/tabs';
 
-import Button from 'components/button';
-import Item from 'components/features/target-spf-item';
-import Loading from 'components/loading';
+import { useSelectedFeatures } from 'hooks/features';
+import { useScenario } from 'hooks/scenarios';
 
-export interface ScenariosFeaturesListProps {
-  onBack: () => void;
+import TargetFeatures from 'layout/scenarios/edit/features/set-up/targets/list';
+
+import Icon from 'components/icon';
+
+import FEATURES_SVG from 'svgs/ui/features.svg?sprite';
+
+export interface ScenariosSidebarEditFeaturesProps {
+
 }
 
-export const ScenariosFeaturesList: React.FC<ScenariosFeaturesListProps> = ({
-  onBack,
-}: ScenariosFeaturesListProps) => {
-  const [submitting, setSubmitting] = useState(false);
+export const ScenariosSidebarEditFeatures: React.FC<ScenariosSidebarEditFeaturesProps> = () => {
   const { query } = useRouter();
   const { sid } = query;
 
-  const selectedFeaturesMutation = useSaveSelectedFeatures({});
-  const saveScenarioMutation = useSaveScenario({
-    requestConfig: {
-      method: 'PATCH',
-    },
-  });
+  const scenarioSlice = getScenarioEditSlice(sid);
+  const { setSubTab } = scenarioSlice.actions;
+
+  const { tab } = useSelector((state) => state[`/scenarios/${sid}/edit`]);
+  const dispatch = useDispatch();
+
+  const { data: scenarioData } = useScenario(sid);
 
   const {
     data: selectedFeaturesData,
   } = useSelectedFeatures(sid, {});
 
-  const {
-    data: targetedFeaturesData,
-    isFetching: targetedFeaturesIsFetching,
-    isFetched: targetedFeaturesIsFetched,
-  } = useTargetedFeatures(sid);
-
-  const { data: scenarioData } = useScenario(sid);
-  const { metadata } = scenarioData || {};
-
-  const INITIAL_VALUES = useMemo(() => {
-    return {
-      features: targetedFeaturesData,
-    };
-  }, [targetedFeaturesData]);
-
-  // Callbacks
-  const onChangeTarget = useCallback((id, v, input) => {
-    const { value, onChange } = input;
-    const features = [...value];
-
-    const feature = features.find((f) => f.id === id);
-    const featureIndex = features.findIndex((f) => f.id === id);
-
-    features[featureIndex] = {
-      ...feature,
-      target: v,
-    };
-    onChange(features);
+  useEffect(() => {
+    // setStep(metaSubtab === ScenarioSidebarSubTabs.FEATURES_TARGET ? 1 : 0);
   }, []);
 
-  const onChangeTargetAll = useCallback((v, input) => {
-    const { value, onChange } = input;
-    const features = [...value].map((f) => ({
-      ...f,
-      target: v,
-    }));
-    onChange(features);
-  }, []);
-
-  const onChangeFPF = useCallback((id, v, input) => {
-    const { value, onChange } = input;
-    const features = [...value];
-
-    const feature = features.find((f) => f.id === id);
-    const featureIndex = features.findIndex((f) => f.id === id);
-
-    features[featureIndex] = {
-      ...feature,
-      fpf: v,
-    };
-    onChange(features);
-  }, []);
-
-  const onChangeFPFAll = useCallback((v, input) => {
-    const { value, onChange } = input;
-    const features = [...value].map((f) => ({
-      ...f,
-      fpf: v,
-    }));
-    onChange(features);
-  }, []);
-
-  const onRemove = useCallback((id, input) => {
-    const { value, onChange } = input;
-    const features = [...value];
-
-    const featureIndex = features.findIndex((f) => f.id === id);
-    features.splice(featureIndex, 1);
-    onChange(features);
-  }, []);
-
-  const onSubmit = useCallback((values) => {
-    setSubmitting(true);
-    const { features } = values;
-
-    const data = {
-      status: 'created',
-      features: selectedFeaturesData.map((sf) => {
-        const { featureId, kind, geoprocessingOperations } = sf;
-
-        if (kind === 'withGeoprocessing') {
-          return {
-            featureId,
-            kind,
-            geoprocessingOperations: geoprocessingOperations.map((go) => {
-              const { splits } = go;
-              console.info(values);
-
-              return {
-                ...go,
-                splits: splits.map((s) => {
-                  const { target, fpf = 1 } = features.find((f) => {
-                    return f.parentId === featureId && f.value === s.value;
-                  });
-
-                  return {
-                    ...s,
-                    marxanSettings: {
-                      prop: target / 100 || 0.5,
-                      fpf,
-                    },
-                  };
-                }),
-              };
-            }),
-
-          };
-        }
-
-        const { target, fpf = 1 } = features.find((f) => f.featureId === featureId);
-        return {
-          featureId,
-          kind,
-          marxanSettings: {
-            prop: target / 100 || 0.5,
-            fpf,
-          },
-        };
-      }),
-    };
-
-    // // Save current features
-    selectedFeaturesMutation.mutate({
-      id: `${sid}`,
-      data,
-    }, {
-      onSuccess: () => {
-        saveScenarioMutation.mutate({
-          id: `${sid}`,
-          data: {
-            metadata: mergeScenarioStatusMetaData(metadata, {
-              tab: ScenarioSidebarTabs.FEATURES,
-              subtab: null,
-            }),
-          },
-        }, {
-          onSuccess: () => {
-            setSubmitting(false);
-          },
-          onError: () => {
-            setSubmitting(false);
-          },
-        });
-      },
-      onError: () => {
-        setSubmitting(false);
-      },
-    });
-  }, [
-    sid,
-    metadata,
-    selectedFeaturesData,
-    selectedFeaturesMutation,
-    saveScenarioMutation,
-  ]);
-
-  // Render
-  if (targetedFeaturesIsFetching && !targetedFeaturesIsFetched) {
-    return (
-      <Loading
-        visible
-        className="z-40 flex items-center justify-center w-full h-40 bg-transparent bg-opacity-90"
-        iconClassName="w-10 h-10 text-primary-500"
-      />
-    );
-  }
+  if (!scenarioData || tab !== ScenarioSidebarTabs.FEATURES) return null;
 
   return (
-    <FormRFF
-      key="features-target"
-      onSubmit={onSubmit}
-      initialValues={INITIAL_VALUES}
-    >
-      {({ handleSubmit, values }) => (
-        <form onSubmit={handleSubmit} autoComplete="off" className="relative flex flex-col flex-grow overflow-hidden">
-          <Loading
-            visible={submitting || targetedFeaturesIsFetching}
-            className="absolute top-0 bottom-0 left-0 right-0 z-40 flex items-center justify-center w-full h-full bg-gray-700 bg-opacity-90"
-            iconClassName="w-10 h-10 text-white"
-          />
-          <p className="text-base opacity-40">
-            Set the minimum targets amount you want to
-            <br />
-            achieve for your features.
-          </p>
-          {(!targetedFeaturesData || !targetedFeaturesData.length) && (
-            <div className="flex items-center justify-center w-full h-40 text-sm uppercase">
-              No results found
+    <div className="flex flex-col flex-grow w-full h-full overflow-hidden">
+      <motion.div
+        key="features"
+        className="flex flex-col min-h-0 overflow-hidden"
+        initial={{ opacity: 0 }}
+        animate={{ opacity: 1 }}
+      >
+
+        <header className="flex items-start justify-between flex-shrink-0">
+          <div>
+            <div className="flex items-baseline space-x-4">
+              <h2 className="text-lg font-medium font-heading">Target and SPF</h2>
+              {/* <InfoButton>
+                <div>
+                  <h4 className="font-heading text-lg mb-2.5">What are features?</h4>
+                  <div className="space-y-2">
+                    <p>
+                      Features are the important habitats, species, processes,
+                      activities, and discrete areas that you want to consider
+                      in your planning process. Common feature data formats are
+                      range maps, polygons, abundances, and continuous scale or
+                      probability of occurrence maps (e.g. 0-1). Features can
+                      include more than just ecological data but also be cultural
+                      and socio-economic areas like community fishing grounds
+                      or traditional-use areas, and other human activities and
+                      industries. Every feature must have a minimum target
+                      amount set.
+                      Some examples include:
+                    </p>
+                    <img src={FEATURE_SPECIES_IMG} alt="Feature-Range" />
+                    <img src={FEATURE_ABUND_IMG} alt="Feature-Abundance" />
+                    <img src={FEATURE_SOCIAL_IMG} alt="Feature-Social" />
+                  </div>
+                </div>
+              </InfoButton> */}
             </div>
-          )}
 
-          {!!targetedFeaturesData && !!targetedFeaturesData.length && (
-            <div className="relative flex flex-col flex-grow overflow-hidden">
-              <div className="absolute top-0 left-0 z-10 w-full h-6 pointer-events-none bg-gradient-to-b from-gray-700 via-gray-700" />
-              <div className="relative h-full px-0.5 overflow-x-visible overflow-y-auto">
-                <FieldRFF name="features">
-                  {({ input }) => (
-                    <div className="py-6">
-                      <Item
-                        id="all-targets"
-                        defaultTarget={50}
-                        defaultFPF={1}
-                        isAllTargets
-                        onChangeTarget={(v) => {
-                          onChangeTargetAll(v, input);
-                        }}
-                        onChangeFPF={(v) => {
-                          onChangeFPFAll(v, input);
-                        }}
-                      />
-
-                      {values.features.map((item, i) => {
-                        return (
-                          <div
-                            className={cx({
-                              'mt-1.5': i !== 0,
-                            })}
-                            key={`${item.id}`}
-                          >
-                            <Item
-                              {...item}
-                              onChangeTarget={(v) => {
-                                onChangeTarget(item.id, v, input);
-                              }}
-                              onChangeFPF={(v) => {
-                                onChangeFPF(item.id, v, input);
-                              }}
-                              onRemove={() => {
-                                onRemove(item.id, input);
-                              }}
-                            />
-                          </div>
-                        );
-                      })}
-                    </div>
-                  )}
-                </FieldRFF>
+            <div className="flex items-center mt-2 space-x-2">
+              <Icon icon={FEATURES_SVG} className="w-4 h-4 text-gray-400" />
+              <div className="text-xs uppercase font-heading">
+                Features:
+                {' '}
+                {selectedFeaturesData && <span className="ml-1 text-gray-400">{selectedFeaturesData.length}</span>}
               </div>
-              <div className="absolute bottom-0 left-0 z-10 w-full h-6 pointer-events-none bg-gradient-to-t from-gray-700 via-gray-700" />
             </div>
-          )}
+          </div>
+        </header>
 
-          {!!targetedFeaturesData && !!targetedFeaturesData.length && (
-            <div className="flex justify-center flex-shrink-0 space-x-3">
-              <Button
-                className="w-full"
-                type="button"
-                theme="secondary"
-                size="lg"
-                onClick={onBack}
-              >
-                Set features
-              </Button>
-
-              <Button
-                className="w-full"
-                type="submit"
-                theme="primary"
-                size="lg"
-                disabled={submitting}
-              >
-                Save
-              </Button>
-            </div>
-          )}
-        </form>
-      )}
-    </FormRFF>
+        <TargetFeatures
+          onBack={() => {
+            dispatch(setSubTab(ScenarioSidebarSubTabs.FEATURES_ADD));
+          }}
+        />
+      </motion.div>
+    </div>
   );
 };
 
-export default ScenariosFeaturesList;
+export default ScenariosSidebarEditFeatures;
